1. int -> Long 타입으로 변경 빠뜨린 부분 수정

+++ src/main/java/com/example/demo/src/store/service/StoreService.java
@@ -27,7 +27,7 @@ public class StoreService {
                 .collect(Collectors.toList());
     }
 
-    public StoreResponse getStoreById(int id) {
+    public StoreResponse getStoreById(Long id) {
         Store store = storeMapper.getStoreById(id);
         if(store == null) {
             // 음식점이 없으면 예외 발생
